RL/reanimator: load/play the path generated by ompl.
This commit also:
1. Use double precision for state tracking and collision detection, to
   match the precision chosed by ompl.
2. Do not include ${CGAL_USE_FILE} in CMakeLists.txt. Now we can enable
   debug build as normal. (INCLUDE(${CGAL_USE_FILE}) apparently would
   enforce Release build type)

+++ lib/osr/cdmodel.h
@@ -9,7 +9,7 @@
 
 namespace osr {
 class Scene;
-using Transform = Eigen::Transform<float, 3, Eigen::AffineCompact>;
+using Transform = Eigen::Transform<double, 3, Eigen::AffineCompact>;
 
 class CDModel {
 	struct CDModelData;
